Requires privilege for deleting error message

+++ html/log/index.php
@@ -67,7 +67,7 @@
             echo "<p>Time: <c id='uptime'></c></p>";
             echo "<p> Temp=<c id='temp'></c>˚C</p>";
             echo "<script>var serverStatus = true;</script>";
-            if (file_exists("../error.log")) {
+            if (file_exists("../error.log") and $PRIVILEGE["deleteError"]) {
                 if ($_POST["delete"] == "error") {
                     unlink("../error.log");
                     writeLog(4, "Error log deleted by $USERNAME or $address");
